暂存

+++ thirteen-authorization/thirteen-authorization-core-service/src/main/java/org/thirteen/authorization/service/impl/base/BaseServiceImpl.java
@@ -7,6 +7,8 @@ import org.springframework.data.domain.ExampleMatcher;
 import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import org.springframework.util.Assert;
+import org.thirteen.authorization.common.utils.StringUtil;
 import org.thirteen.authorization.dozer.DozerMapper;
 import org.thirteen.authorization.exceptions.DataNotFoundException;
 import org.thirteen.authorization.model.po.base.BasePO;
@@ -17,9 +19,12 @@ import org.thirteen.authorization.service.support.base.ModelInformation;
 import org.thirteen.authorization.service.support.base.ModelSupport;
 
 import java.lang.reflect.Method;
+import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Optional;
 import java.util.stream.Collectors;
+
+import static org.thirteen.authorization.service.support.base.ModelSupport.*;
 
 /**
  * @author Aaron.Sun
@@ -30,15 +35,6 @@ import java.util.stream.Collectors;
 @Service
 public abstract class BaseServiceImpl<VO extends BaseVO<PK>, PK, PO extends BasePO<PK>> implements BaseService<VO, PK> {
 
-    /**
-     * 逻辑删除字段
-     */
-    public static final String DEL_FLAG_FIELD = "delFlag";
-    /**
-     * 删除标记（0：正常；1：删除）
-     */
-    public static final String DEL_FLAG_NORMAL = "0";
-    public static final String DEL_FLAG_DELETE = "1";
     /**
      * baseRepository注入
      */
@@ -86,37 +82,46 @@ public abstract class BaseServiceImpl<VO extends BaseVO<PK>, PK, PO extends Base
     @Transactional(rollbackFor = Exception.class)
     @Override
     public void save(VO model) {
-        baseRepository.save(dozerMapper.map(model, poClass));
+        Assert.notNull(model, "Entity must not be null!");
+        PO po = dozerMapper.map(model, poClass);
+        baseRepository.save(poSupport.getModel(po, CREATE_BY_FIELD, CREATE_TIME_FIELD, true));
     }
 
     @Transactional(rollbackFor = Exception.class)
     @Override
     public void saveAndFlush(VO model) {
-        baseRepository.saveAndFlush(dozerMapper.map(model, poClass));
+        Assert.notNull(model, "Entity must not be null!");
+        PO po = dozerMapper.map(model, poClass);
+        baseRepository.saveAndFlush(poSupport.getModel(po, CREATE_BY_FIELD, CREATE_TIME_FIELD, true));
     }
 
     @Transactional(rollbackFor = Exception.class)
     @Override
     public void saveAll(List<VO> models) {
-        baseRepository.saveAll(dozerMapper.mapList(models, poClass));
+        Assert.notEmpty(models, "Entity collection must not be empty!");
+        List<PO> pos = dozerMapper.mapList(models, poClass);
+        baseRepository.saveAll(poSupport.getModels(pos, CREATE_BY_FIELD, CREATE_TIME_FIELD, true));
     }
 
     @Transactional(rollbackFor = Exception.class)
     @Override
     public void update(VO model) {
-        baseRepository.save(dozerMapper.map(model, poClass));
+        Assert.notNull(model, "Entity must not be null!");
+        PO po = dozerMapper.map(model, poClass);
+        baseRepository.save(poSupport.getModel(po, UPDATE_BY_FIELD, UPDATE_TIME_FIELD, false));
     }
 
     @Transactional(rollbackFor = {Exception.class})
     @Override
     public void updateSelective(VO model) {
+        Assert.notNull(model, "Entity must not be null!");
         // 更新前先由ID获取数据信息
         Optional<PO> optional = baseRepository.findById(model.getId());
         // 判断数据是否存在
         if (optional.isPresent()) {
             // 如果存在则将入参中的非null属性赋给查询结果
             dozerMapper.copy(dozerMapper.map(model, poClass), optional.get(), false, true);
-            baseRepository.save(optional.get());
+            baseRepository.save(poSupport.getModel(optional.get(), UPDATE_BY_FIELD, UPDATE_TIME_FIELD, false));
         } else {
             throw new DataNotFoundException(model.getId());
         }
@@ -125,60 +130,71 @@ public abstract class BaseServiceImpl<VO extends BaseVO<PK>, PK, PO extends Base
     @Transactional(rollbackFor = Exception.class)
     @Override
     public void updateAndFlush(VO model) {
-        baseRepository.saveAndFlush(dozerMapper.map(model, poClass));
+        Assert.notNull(model, "Entity must not be null!");
+        PO po = dozerMapper.map(model, poClass);
+        baseRepository.saveAndFlush(poSupport.getModel(po, UPDATE_BY_FIELD, UPDATE_TIME_FIELD, false));
     }
 
     @Transactional(rollbackFor = Exception.class)
     @Override
     public void updateAll(List<VO> models) {
-        baseRepository.saveAll(dozerMapper.mapList(models, poClass));
+        Assert.notEmpty(models, "Entity collection must not be empty!");
+        List<PO> pos = dozerMapper.mapList(models, poClass);
+        baseRepository.saveAll(poSupport.getModels(pos, UPDATE_BY_FIELD, UPDATE_TIME_FIELD, false));
     }
 
     @Transactional(rollbackFor = Exception.class)
     @Override
     public void delete(PK id) {
-        baseRepository.deleteById(id);
+        Assert.notNull(id, "The given id must not be null!");
+        // 如果存在逻辑删除字段，则逻辑删除
+        if (poInformation.contains(DEL_FLAG_FIELD)) {
+            // 获取poClass中逻辑删除字段的set方法，获取不到则抛出异常
+            Method method = poInformation.getSetter(DEL_FLAG_FIELD, String.class);
+            // 逻辑删除前先由ID获取数据信息
+            Optional<PO> optional = baseRepository.findById(id);
+            // 判断数据是否存在
+            if (optional.isPresent()) {
+                // 如果存在则将入参中的非null属性赋给查询结果
+                poInformation.invokeSet(method, optional.get(), DEL_FLAG_DELETE);
+                baseRepository.save(optional.get());
+            } else {
+                throw new DataNotFoundException(id);
+            }
+        } else {
+            baseRepository.deleteById(id);
+        }
     }
 
     @Transactional(rollbackFor = Exception.class)
     @Override
     public void deleteAll(List<PK> ids) {
-        baseRepository.deleteAll(ids.stream().map(item -> poSupport.builder().id(item).build())
-            .collect(Collectors.toList()));
+        Assert.notEmpty(ids, "ID collection must not be empty!");
+        // 如果存在逻辑删除字段，则逻辑删除
+        if (poInformation.contains(DEL_FLAG_FIELD)) {
+            ids.forEach(this::delete);
+        } else {
+            baseRepository.deleteAll(ids.stream().map(item -> poSupport.builder().id(item).build())
+                .collect(Collectors.toList()));
+        }
     }
 
     @Transactional(rollbackFor = Exception.class)
     @Override
     public void deleteInBatch(List<PK> ids) {
-        baseRepository.deleteInBatch(ids.stream().map(item -> poSupport.builder().id(item).build())
-            .collect(Collectors.toList()));
-    }
-
-    @Transactional(rollbackFor = {Exception.class})
-    @Override
-    public void logicDelete(PK id) {
-        // 获取poClass中逻辑删除字段的set方法，获取不到则抛出异常
-        Method method = poInformation.getSetter(DEL_FLAG_FIELD, String.class);
-        // 逻辑删除前先由ID获取数据信息
-        Optional<PO> optional = baseRepository.findById(id);
-        // 判断数据是否存在
-        if (optional.isPresent()) {
-            // 如果存在则将入参中的非null属性赋给查询结果
-            poInformation.invokeSet(method, optional.get(), DEL_FLAG_DELETE);
-            baseRepository.save(optional.get());
+        Assert.notEmpty(ids, "ID collection must not be empty!");
+        // 如果存在逻辑删除字段，则逻辑删除
+        if (poInformation.contains(DEL_FLAG_FIELD)) {
+            ids.forEach(this::delete);
         } else {
-            throw new DataNotFoundException(id);
+            baseRepository.deleteInBatch(ids.stream().map(item -> poSupport.builder().id(item).build())
+                .collect(Collectors.toList()));
         }
-    }
-
-    @Transactional(rollbackFor = {Exception.class})
-    @Override
-    public void logicDeleteAll(List<PK> ids) {
-        ids.forEach(this::logicDelete);
     }
 
     @Override
     public VO get(PK id) {
+        Assert.notNull(id, "The given id must not be null!");
         Optional<PO> optional = baseRepository.findById(id);
         return optional.map(po -> dozerMapper.map(po, voClass)).orElse(null);
     }
@@ -198,7 +214,4 @@ public abstract class BaseServiceImpl<VO extends BaseVO<PK>, PK, PO extends Base
         Example<PO> example = Example.of(poSupport.builder().build(), matcher);
         return dozerMapper.mapList(baseRepository.findAll(example), voClass);
     }
-
-
-
 }
